+++ packages/openapi_base/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.1
+
+* bump UUID version constraint to support `4.x`
+
+## 1.3.0
+
+* `ResponseMap` now has a return type.
+
 ## 1.1.1
 
 * string binary support