+++ packages/openapi_code_builder/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.3.0
+
+* Update SDK Constraint to `>= 3.2.0 <4.0.0`
+* add `onElse` to `map` functions and allow a return value.
+* Make base Response classes `sealed`.
+
+## 1.2.0
+
+* Support `DateTime` format for parameters.
+ 
 ## 1.1.4+5
 
 * string binary support